requirements.txt
flask==3.0.0
google-cloud-bigquery==3.14.1
python-dotenv==1.0.0
